pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = "This project is an API for an e-commerce system, built with Django Rest Framework (DRF). The API will provide endpoints to manage products, categories, shopping cart and orders, with a focus on simplicity, scalability and easy integration with frontends and external systems."
authors = ["Gabriel Nathan <devgabrielnathan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
django = "^5.1.4"
python-decouple = "^3.8"
dj-database-url = "^2.3.0"
django-extensions = "^3.2.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.3"
taskipy = "^1.14.1"

[tool.taskipy.tasks]
base = 'python manage.py'
help = 'python manage.py help'
run = 'python manage.py runserver'
app = 'python manage.py startapp'
adm = 'python manage.py createsuperuser'
mg = 'python manage.py migrate'
mk = 'python manage.py makemigrations'
test = 'python manage.py test'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[tool.ruff]
line-length = 79
indent-width = 4
extend-exclude = ['migrations']
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "settings.py",
    "manage.py",
]

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ["E501", "PLR6301", "PT009"]

[tool.ruff.format]
preview = true
quote-style = 'single'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
